src/knowledge/pyproject.toml
[project]
name = "knowledge"
version = "0.1.0"
description = ""
authors = [
    {name = "Artur Drohomeretskyi",email = "artur.drohomeretskyi@hys-enterprise.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
package-mode = false
dependencies = [
    "langchain-text-splitters (>=0.3.6,<0.4.0)",
    "openai (>=1.65.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
